Refactor WalletConnect and remove unused code
